fix: error phpcs: pemisahan dua class
