+++ basics/14_oop_dasar/6_constant_magic_constants2b.php
@@ -1,25 +1,6 @@
 <?php
 
-/**
- * Class Pengguna
- */
-class Pengguna
-{
-    // Deklarasi konstanta
-    public const PESAN_LOGIN = "Selamat datang‼️";
-    public const PESAN_LOGOUT = "Sampai jumpa lagi 👋🏻";
-
-    // Method login dan logout
-    public function login()
-    {
-        echo self::PESAN_LOGIN . "\n";
-    }
-
-    public function logout()
-    {
-        echo self::PESAN_LOGOUT . "\n";
-    }
-}
+namespace Konstanta;
 
 /**
  * Class Lingkaran
